stereo audio bug fix
VAD and other parts of engine work on mono audio.
Now, stereo audio is converted to mono audio by taking mean of both the channels in remake_audio().

Also, silero-vad works on 8000 or 16000Hz sample rate. So, if sample rate passed is not a multiple of 16000, we set it to 16000 in vad()