create streamlit application
main.py contains the streamlit application that uses the pretrained model to generate text from the recorded audio or uploaded video successfully in both the cases.
However, when recording audio, it is imperative to pronounce words in clear fashion and with precision, as the model is trained on a very small dataset, which lacks variety of accents.

To run the application, head to terminal and open the current path to repository, and run "streamlit run main.py". This will launch a browser window with the application.